+++ src/app/(providers)/(styles)/challenge/create/_components/Milestone/MilestoneComponent.tsx
@@ -5,6 +5,7 @@ import { nanoid } from "nanoid"
 
 import Input from "@/components/Input"
 
+import MilestoneComponetMobile from "./MilestoneComponetMobile"
 import RoutineComponent from "./RoutineComponent"
 
 type MilestoneType = {
@@ -60,67 +61,69 @@ function MilestoneComponent({
     })
   }
   return (
-    <div
-      ref={provided.innerRef}
-      {...provided.draggableProps}
-      {...provided.dragHandleProps}
-      className="mb-4 flex flex-col gap-2 rounded-lg border border-gray-400 bg-gray-100 p-4"
-    >
-      <button onClick={() => deleteMilestone(milestone.id)}>삭제</button>
-      {/* 각 마일스톤 내 루틴 Droppable */}
-      <Droppable droppableId={milestone.id} type="routine">
-        {(provided) => (
-          <div
-            {...provided.droppableProps}
-            ref={provided.innerRef}
-            className="flex h-full flex-col justify-between p-4"
-          >
-            <h2 className="mb-2 text-lg font-semibold">
-              마일스톤: {milestone.name}
-            </h2>
-            {milestone.routines.map((routine, index) => (
-              <Draggable
-                key={routine.id}
-                draggableId={routine.id}
-                index={index}
-              >
-                {(provided, snapshot) => (
-                  <RoutineComponent
-                    provided={provided}
-                    snapshot={snapshot}
-                    routine={routine}
-                    milestoneId={milestone.id}
-                    deleteRoutine={deleteRoutine}
-                  />
-                )}
-              </Draggable>
-            ))}
-            {provided.placeholder}
-            <div className="mt-auto flex flex-col gap-2">
-              <Input
-                label="루틴 생성"
-                value={routineValue}
-                onChange={(e) => {
-                  setRoutineValue(e.target.value)
-                }}
-              />
-              <button
-                onClick={() => {
-                  setRoutineValue("")
-                  createRoutine(milestone.id, {
-                    id: nanoid(),
-                    content: routineValue,
-                  })
-                }}
-                className="flex h-[45px] items-center justify-center rounded border"
-              >
-                생성하기
-              </button>
+    <>
+      <div
+        ref={provided.innerRef}
+        {...provided.draggableProps}
+        {...provided.dragHandleProps}
+        className="mb-4 hidden flex-col gap-2 rounded-lg border border-gray-400 bg-gray-100 p-4 sm:flex"
+      >
+        <button onClick={() => deleteMilestone(milestone.id)}>삭제</button>
+        {/* 각 마일스톤 내 루틴 Droppable */}
+        <Droppable droppableId={milestone.id} type="routine">
+          {(provided) => (
+            <div
+              {...provided.droppableProps}
+              ref={provided.innerRef}
+              className="flex h-full flex-col justify-between p-4"
+            >
+              <h2 className="mb-2 text-lg font-semibold">
+                마일스톤: {milestone.name}
+              </h2>
+              {milestone.routines.map((routine, index) => (
+                <Draggable
+                  key={routine.id}
+                  draggableId={routine.id}
+                  index={index}
+                >
+                  {(provided, snapshot) => (
+                    <RoutineComponent
+                      provided={provided}
+                      snapshot={snapshot}
+                      routine={routine}
+                      milestoneId={milestone.id}
+                      deleteRoutine={deleteRoutine}
+                    />
+                  )}
+                </Draggable>
+              ))}
+              <div className="mt-auto flex flex-col gap-2">
+                <Input
+                  label="루틴 생성"
+                  value={routineValue}
+                  onChange={(e) => {
+                    setRoutineValue(e.target.value)
+                  }}
+                />
+                <button
+                  onClick={() => {
+                    setRoutineValue("")
+                    createRoutine(milestone.id, {
+                      id: nanoid(),
+                      content: routineValue,
+                    })
+                  }}
+                  className="flex h-[45px] items-center justify-center rounded border"
+                >
+                  생성하기
+                </button>
+              </div>
             </div>
-          </div>
-        )}
-      </Droppable>
-    </div>
+          )}
+        </Droppable>
+      </div>
+      <MilestoneComponetMobile provided={provided} milestone={milestone} />
+    </>
   )
 }
